Moved Game to the main package and renamed Board to GameController
